\name{EBSC-internal}

\alias{addTrans}
\alias{print.ebssc}
\alias{ebssc.default}
\alias{ebsscEst}
\alias{EBCparallel}
\alias{EBCq}
\alias{eBsc.default}
\alias{eigenvalues}
\alias{Elambda}
\alias{Eq}
\alias{get_lambda}
\alias{get_q}
\alias{lmm}
\alias{mycorMatrix}
\alias{print.eBsc}
\alias{Qfunc}
\alias{smoothrhos}
\alias{Turn}

\title{Internal ebssc objects}
\description{Internal functions}
\details{These functions are not to be called by the user}
\keyword{internal}
